used form validatiors to make sure no input is left empty

+++ app/main/forms.py
@@ -1,12 +1,12 @@
 
 from flask_wtf import FlaskForm
-from wtforms import IntegerField, SelectField, SubmitField
+from wtforms import FloatField, SelectField, SubmitField
 from wtforms.validators import DataRequired
 
 
 class CalculatorForm(FlaskForm):
-    num1 = IntegerField('Enter First Number', validators=[DataRequired()])
-    num2 = IntegerField('Enter Second Number', validators=[DataRequired()])
+    num1 = FloatField('Enter First Number', validators=[DataRequired()])
+    num2 = FloatField('Enter Second Number', validators=[DataRequired()])
     operator = SelectField('Operator',
                            choices=[('Multiply', 'Multiply: x'), ('Add', 'Add: +'), ('Divide',
                                                                                      'Divide: /'), ('Subtract', 'Subtract: -'), ('Reminder', 'Reminder: %')]
